% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.sum.R
\name{evaluation.sum}
\alias{evaluation.sum}
\title{Hierarchical Heterogeneity Regression Analysis.}
\usage{
evaluation.sum(n,q,p,admmres, abic.n, admmres2, Beta0, bic.var)
}
\arguments{
\item{n}{The sample size.}

\item{q}{The dimension of type 1 features.}

\item{p}{The dimension of type 2 features.}

\item{admmres}{The results corresponding to lambda1.}

\item{abic.n}{The BIC values.}

\item{admmres2}{The results corresponding to lambda1.}

\item{Beta0}{The true values of beta.}

\item{bic.var}{The BIC values.}
}
\value{
A result list including: evaluating indicator
}
\description{
The main function for Transfer learning for tensor graphical models.
}
