% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.GGM.R
\docType{data}
\name{example.data.GGM}
\alias{example.data.GGM}
\title{Some example data}
\format{
A list.
}
\source{
Simulated data
}
\description{
Some example data
}
\examples{
data(example.data.GGM)
}
